--- src/students/offeredCourse/course.entity.ts
@@ -1,20 +1,0 @@
-import { IsNotEmpty } from "class-validator";
-import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
-
-@Entity("courses")
-export class OfferedCoursesEntity{
-  @PrimaryGeneratedColumn()
-  id: string;
-
-  @Column()
-  @IsNotEmpty()
-  courseCode: string;
-
-  @Column()
-  @IsNotEmpty()
-  courseName: string;
-
-  @Column()
-  @IsNotEmpty()
-  semister: string;
-}
